requirement.txt
langgraph==0.2.28 
langchain-openai==0.2.2 
langchain==0.3.12
requests==2.32.3
python-dotenv==1.0.0
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0